+++ apps/playground/src/app/array-values/page.tsx
@@ -1,12 +1,18 @@
 "use client";
 
+import { useState } from "react";
+import { useSearchParams } from "next/navigation";
 import { SetKeyValueArrayInputs } from "@/components/set-key-value-inputs";
 import { Alert } from "@/components/ui/alert";
 
-import { useSearchParamsState } from "@use-search-params-state/next";
+import { searchParamsToObject, useObserveAndStore } from "@sp-hooks/next";
 
 export default function Page() {
-  const [state, setState] = useSearchParamsState();
+  const sp = useSearchParams();
+
+  const [state, setState] = useState(searchParamsToObject(sp));
+
+  useObserveAndStore(state);
 
   return (
     <div>
@@ -14,7 +20,11 @@ export default function Page() {
         {"This demonstration shows how to use the library with arrays."}
       </Alert>
 
-      <SetKeyValueArrayInputs setState={(k, v) => setState(k, v)} />
+      <SetKeyValueArrayInputs
+        setState={(key, values) => {
+          setState((s) => ({ ...s, [key]: values }));
+        }}
+      />
 
       <pre>{JSON.stringify(state, null, 2)}</pre>
     </div>