+++ apps/playground/src/app/basic/page.tsx
@@ -1,11 +1,17 @@
 "use client";
 
+import { useState } from "react";
+import { useSearchParams } from "next/navigation";
 import { SetKeyValueInputs } from "@/components/set-key-value-inputs";
 
-import { useSearchParamsState } from "@use-search-params-state/next";
+import { searchParamsToObject, useObserveAndStore } from "@sp-hooks/next";
 
 export default function Page() {
-  const [state, setState] = useSearchParamsState();
+  const sp = useSearchParams();
+
+  const [state, setState] = useState(searchParamsToObject(sp));
+
+  useObserveAndStore(state);
 
   return (
     <div>