+++ apps/playground/src/app/layout.tsx
@@ -2,10 +2,10 @@ import { Inter as FontSans } from "next/font/google";
 
 import "./globals.css";
 
+import { Suspense } from "react";
 import { Analytics } from "@/components/analytics";
 import { ThemeProvider } from "@/components/theme-provider";
 import { Toaster } from "@/components/ui/toaster";
-import { siteConfig } from "@/config/site";
 import { cn } from "@/lib/utils";
 
 import { Sidebar } from "./sidebar";
@@ -18,53 +18,6 @@ const fontSans = FontSans({
 interface RootLayoutProps {
   children: React.ReactNode;
 }
-
-export const metadata = {
-  title: {
-    default: siteConfig.name,
-    template: `%s | ${siteConfig.name}`,
-  },
-  description: siteConfig.description,
-  keywords: [
-    "Next.js",
-    "React",
-    "Tailwind CSS",
-    "Server Components",
-    "Radix UI",
-  ],
-  authors: [
-    {
-      name: "shadcn",
-      url: "https://shadcn.com",
-    },
-  ],
-  creator: "shadcn",
-  themeColor: [
-    { media: "(prefers-color-scheme: light)", color: "white" },
-    { media: "(prefers-color-scheme: dark)", color: "black" },
-  ],
-  openGraph: {
-    type: "website",
-    locale: "en_US",
-    url: siteConfig.url,
-    title: siteConfig.name,
-    description: siteConfig.description,
-    siteName: siteConfig.name,
-  },
-  twitter: {
-    card: "summary_large_image",
-    title: siteConfig.name,
-    description: siteConfig.description,
-    images: [`${siteConfig.url}/og.jpg`],
-    creator: "@shadcn",
-  },
-  icons: {
-    icon: "/favicon.ico",
-    shortcut: "/favicon-16x16.png",
-    apple: "/apple-touch-icon.png",
-  },
-  manifest: `${siteConfig.url}/site.webmanifest`,
-};
 
 export default function RootLayout({ children }: RootLayoutProps) {
   return (
@@ -79,7 +32,9 @@ export default function RootLayout({ children }: RootLayoutProps) {
         <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
           <main>
             <Sidebar />
-            <div className="ml-64 p-4">{children}</div>
+            <div className="ml-64 p-4">
+              <Suspense fallback={null}>{children}</Suspense>
+            </div>
           </main>
           <Analytics />
           <Toaster />