+++ packages/react/src/utils.ts
@@ -1,63 +1,18 @@
+import { useEffect } from "react";
 import type { ReadonlyURLSearchParams } from "next/navigation";
 
-import type { UseSearchParamsStateParams } from "./types";
+import { SPHooksStateType, SupportedValues } from "./types";
 
-export const getSearchParams = <
-  S extends Partial<Record<string, string | string[]>>,
->({
-  newObject,
-  options,
-  initiallySetKeys,
-}: {
-  newObject: S;
-  options: UseSearchParamsStateParams<Record<string, unknown>>;
-  initiallySetKeys: string[];
-}) => {
-  const sp = new URLSearchParams();
-
-  Object.keys(newObject).forEach((k) => {
-    const v = newObject[k];
-    if (typeof v === "undefined") return;
-
-    // If the key was set initially, we want to set save it regardless of value.
-    if (options.preserveInitialKeys && initiallySetKeys.includes(k)) {
-      if (Array.isArray(v)) {
-        v.forEach((value) => {
-          sp.append(k, value);
-        });
-      } else {
-        sp.set(k, v);
-      }
-
-      return;
-    }
-
-    // If the value is falsy and we want to remove falsy values, skip it.
-    if (options.removeFalsyValues && !v) return;
-
-    // If the value is the default value and we want to remove default values, skip it.
-    if (options.removeDefaultValues && v === options.defaultValues?.[k]) return;
-
-    if (Array.isArray(v)) {
-      v.forEach((value) => {
-        sp.append(k, value);
-      });
-    } else {
-      sp.set(k, v);
-    }
-  });
-
-  if (options.sortKeys) {
-    sp.sort();
-  }
-
-  return sp;
-};
-
-export const searchParamsToObject = (
+/*
+ * Conerts a URLSearchParams object to a plain object, with
+ * the values being either a string or an array of strings,
+ * depending on the number of values for a given key.
+ */
+export const searchParamsToObject = <S extends SPHooksStateType>(
   sp: URLSearchParams | ReadonlyURLSearchParams,
-): Record<string, string | string[]> => {
-  const newObj: Record<string, string | string[]> = {};
+  options?: { defaultValues?: Partial<S> },
+): SPHooksStateType => {
+  const newObj: SPHooksStateType = {};
 
   for (const key of Array.from(sp.keys())) {
     const values = sp.getAll(key);
@@ -71,5 +26,67 @@ export const searchParamsToObject = (
     }
   }
 
+  if (options?.defaultValues) {
+    Object.entries(options.defaultValues).forEach(
+      ([key, value]: [string, string | string[]]) => {
+        if (typeof newObj[key] === "undefined") {
+          newObj[key] = value;
+        }
+      },
+    );
+  }
+
   return newObj;
 };
+
+/*
+ * Converts a plain object to a URLSearchParams object.
+ */
+export function stateToSearchParams<S extends SPHooksStateType>(
+  state: S,
+  options?: { defaultValues?: Partial<S> },
+) {
+  const searchParams = new URLSearchParams();
+
+  Object.entries(state).forEach(([key, value]) =>
+    mutSetValueToSearchParams(searchParams, key, value, options),
+  );
+
+  return searchParams;
+}
+
+/*
+ * Observe a state object and call setSearchParams when the state changes.
+ */
+export function useObserveAndStore<S extends SPHooksStateType>(
+  state: S,
+  setSearchParams: (newSearchParams: URLSearchParams) => void,
+  options?: { defaultValues?: Partial<S> }, // TODO: Change this to ObserveAndStoreOptions
+) {
+  useEffect(() => {
+    setSearchParams(stateToSearchParams(state, options));
+  }, [state]);
+}
+
+/*
+ * Accepts URLSearchParams as first parameter and a generic serialisable type as a second.
+ * Converts type to a string or an array of strings depending on the provided type and sets it in the URLSearchParams.
+ */
+export function mutSetValueToSearchParams<S extends SPHooksStateType>(
+  sp: URLSearchParams,
+  key: string,
+  value: SupportedValues,
+  options?: { defaultValues?: Partial<S> },
+) {
+  if (Array.isArray(value)) {
+    value.forEach((v) => {
+      sp.append(key, encodeURIComponent(v.toString()));
+    });
+  } else {
+    if (options?.defaultValues && options.defaultValues[key] === value) {
+      return;
+    }
+
+    sp.set(key, encodeURIComponent(value.toString()));
+  }
+}