+++ tooling/prettier/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @use-search-params-state/prettier-config
+# @sp-hooks/prettier-config
 
 ## 0.1.1
 
 ### Patch Changes
 
-- c81587f: use-search-params-state mvp release
+- c81587f: sp-hooks mvp release